requirements.txt
fastapi
uvicorn
groq
chromadb
sentence-transformers
python-dotenv
pydantic
httpx
duckduckgo-search
pypdf
python-multipart
